+++ src/misc.f
@@ -553,7 +553,7 @@ c----------------------------------------------------------------------
 
         real*8 grad_phi1_sq
 
-        integer dimA,dimB
+        real*8 dimA,dimB
 
         !--------------------------------------------------------------
         ! variables for tensor manipulations 
@@ -648,8 +648,8 @@ c----------------------------------------------------------------------
         y0=L/2  !NOTE: change this to y0=y(j) when we add y-dependence
 
         ! set dimensions of S3 and S4 subspaces
-        dimA=3
-        dimB=4
+        dimA=3d0 
+        dimB=4d0
 
         ! set gads values using sin(theta1)=sin(theta2)=1 w.l.o.g 
         !(considering theta1,theta2-independent case, so theta1=theta2=pi/2 slice will do)
@@ -1572,15 +1572,15 @@ c----------------------------------------------------------------------
      &     +g0_uu(3,1)*gamma_ull(3,1,3)*gA_x(3)/2 
      &     +g0_uu(3,2)*gamma_ull(3,2,3)*gA_x(3)/2 
      &     +g0_uu(3,3)*gamma_ull(3,3,3)*gA_x(3)/2 
-     &     -g0_uu(1,1)*(gA_x(1)*gA_x(1))/(4*gA)*(dimA-2)
-     &     -g0_uu(1,2)*(gA_x(1)*gA_x(2))/(4*gA)*(dimA-2)
-     &     -g0_uu(1,3)*(gA_x(1)*gA_x(3))/(4*gA)*(dimA-2)
-     &     -g0_uu(2,1)*(gA_x(2)*gA_x(1))/(4*gA)*(dimA-2)
-     &     -g0_uu(2,2)*(gA_x(2)*gA_x(2))/(4*gA)*(dimA-2)
-     &     -g0_uu(2,3)*(gA_x(2)*gA_x(3))/(4*gA)*(dimA-2)
-     &     -g0_uu(3,1)*(gA_x(3)*gA_x(1))/(4*gA)*(dimA-2)
-     &     -g0_uu(3,2)*(gA_x(3)*gA_x(2))/(4*gA)*(dimA-2)
-     &     -g0_uu(3,3)*(gA_x(3)*gA_x(3))/(4*gA)*(dimA-2)
+     &     -g0_uu(1,1)*(gA_x(1)*gA_x(1))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(1,2)*(gA_x(1)*gA_x(2))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(1,3)*(gA_x(1)*gA_x(3))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(2,1)*(gA_x(2)*gA_x(1))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(2,2)*(gA_x(2)*gA_x(2))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(2,3)*(gA_x(2)*gA_x(3))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(3,1)*(gA_x(3)*gA_x(1))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(3,2)*(gA_x(3)*gA_x(2))/(4*gA)*(dimA-2d0)
+     &     -g0_uu(3,3)*(gA_x(3)*gA_x(3))/(4*gA)*(dimA-2d0)
      &     -g0_uu(1,1)*(gA_x(1)*gB_x(1))/(4*gB)*dimB
      &     -g0_uu(1,2)*(gA_x(1)*gB_x(2))/(4*gB)*dimB
      &     -g0_uu(1,3)*(gA_x(1)*gB_x(3))/(4*gB)*dimB
@@ -1627,15 +1627,15 @@ c----------------------------------------------------------------------
      &     +g0_uu(3,1)*gamma_ull(3,1,3)*gB_x(3)/2 
      &     +g0_uu(3,2)*gamma_ull(3,2,3)*gB_x(3)/2 
      &     +g0_uu(3,3)*gamma_ull(3,3,3)*gB_x(3)/2 
-     &     -g0_uu(1,1)*(gB_x(1)*gB_x(1))/(4*gB)*(dimB-2)
-     &     -g0_uu(1,2)*(gB_x(1)*gB_x(2))/(4*gB)*(dimB-2)
-     &     -g0_uu(1,3)*(gB_x(1)*gB_x(3))/(4*gB)*(dimB-2)
-     &     -g0_uu(2,1)*(gB_x(2)*gB_x(1))/(4*gB)*(dimB-2)
-     &     -g0_uu(2,2)*(gB_x(2)*gB_x(2))/(4*gB)*(dimB-2)
-     &     -g0_uu(2,3)*(gB_x(2)*gB_x(3))/(4*gB)*(dimB-2)
-     &     -g0_uu(3,1)*(gB_x(3)*gB_x(1))/(4*gB)*(dimB-2)
-     &     -g0_uu(3,2)*(gB_x(3)*gB_x(2))/(4*gB)*(dimB-2)
-     &     -g0_uu(3,3)*(gB_x(3)*gB_x(3))/(4*gB)*(dimB-2)
+     &     -g0_uu(1,1)*(gB_x(1)*gB_x(1))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(1,2)*(gB_x(1)*gB_x(2))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(1,3)*(gB_x(1)*gB_x(3))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(2,1)*(gB_x(2)*gB_x(1))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(2,2)*(gB_x(2)*gB_x(2))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(2,3)*(gB_x(2)*gB_x(3))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(3,1)*(gB_x(3)*gB_x(1))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(3,2)*(gB_x(3)*gB_x(2))/(4*gB)*(dimB-2d0)
+     &     -g0_uu(3,3)*(gB_x(3)*gB_x(3))/(4*gB)*(dimB-2d0)
      &     -g0_uu(1,1)*(gA_x(1)*gB_x(1))/(4*gA)*dimA 
      &     -g0_uu(1,2)*(gA_x(1)*gB_x(2))/(4*gA)*dimA
      &     -g0_uu(1,3)*(gA_x(1)*gB_x(3))/(4*gA)*dimA
